Move `themes` to `core` feauture
