+++ app/src/main/java/ke/co/xently/shopping/features/core/ui/AutoCompleteTextField.kt
@@ -24,8 +24,8 @@ import androidx.compose.ui.tooling.preview.Preview
 import androidx.compose.ui.unit.dp
 import ke.co.xently.shopping.features.core.hasEmojis
 import ke.co.xently.shopping.features.core.ui.components.AutoCompleteSearchResults
+import ke.co.xently.shopping.features.core.ui.theme.XentlyTheme
 import ke.co.xently.shopping.remotedatasource.services.AutoCompleteService
-import ke.co.xently.shopping.ui.theme.XentlyTheme
 import kotlinx.coroutines.delay
 import kotlin.random.Random
 import kotlin.time.Duration.Companion.milliseconds
