+++ app/src/main/java/ke/co/xently/shopping/features/core/ui/theme/Type.kt
@@ -1,4 +1,4 @@
-package ke.co.xently.shopping.ui.theme
+package ke.co.xently.shopping.features.core.ui.theme
 
 import androidx.compose.material3.Typography
 import androidx.compose.ui.text.TextStyle
