+++ app/src/main/java/ke/co/xently/shopping/features/measurementunit/ui/AddMeasurementUnitPage.kt
@@ -29,10 +29,10 @@ import ke.co.xently.shopping.features.core.hasEmojis
 import ke.co.xently.shopping.features.core.ui.LabeledCheckbox
 import ke.co.xently.shopping.features.core.ui.MultiStepScreen
 import ke.co.xently.shopping.features.core.ui.rememberAutoCompleteTextFieldState
+import ke.co.xently.shopping.features.core.ui.theme.XentlyTheme
 import ke.co.xently.shopping.features.measurementunit.models.MeasurementUnit
 import ke.co.xently.shopping.features.products.models.Product
 import ke.co.xently.shopping.features.products.ui.components.AddProductAutoCompleteTextField
-import ke.co.xently.shopping.ui.theme.XentlyTheme
 
 @Composable
 fun AddMeasurementUnitPage(
