+++ app/src/main/java/ke/co/xently/shopping/features/products/ui/subscreens/AddProductNamePage.kt
@@ -26,10 +26,10 @@ import ke.co.xently.shopping.features.core.ui.LabeledCheckbox
 import ke.co.xently.shopping.features.core.ui.MultiStepScreen
 import ke.co.xently.shopping.features.core.ui.UIState
 import ke.co.xently.shopping.features.core.ui.rememberAutoCompleteTextFieldState
+import ke.co.xently.shopping.features.core.ui.theme.XentlyTheme
 import ke.co.xently.shopping.features.products.models.Product
 import ke.co.xently.shopping.features.products.ui.LocalProductAutoCompleteService
 import ke.co.xently.shopping.features.products.ui.components.AddProductAutoCompleteTextField
-import ke.co.xently.shopping.ui.theme.XentlyTheme
 
 internal sealed class ProductNameUIState(message: Int) : UIState(message) {
     object OK : ProductNameUIState(android.R.string.ok)
