+++ app/src/main/java/ke/co/xently/shopping/features/recommendations/ui/RecommendationResponseDetailsScreen.kt
@@ -34,12 +34,12 @@ import androidx.compose.ui.text.style.TextDecoration
 import androidx.compose.ui.tooling.preview.Preview
 import androidx.compose.ui.unit.dp
 import ke.co.xently.shopping.R
+import ke.co.xently.shopping.features.core.ui.theme.XentlyTheme
 import ke.co.xently.shopping.features.recommendations.models.Recommendation
 import ke.co.xently.shopping.features.recommendations.ui.components.HitItem
 import ke.co.xently.shopping.features.recommendations.ui.components.MissItem
 import ke.co.xently.shopping.features.recommendations.ui.components.RecommendationSummaryItemDropdownMenu
 import ke.co.xently.shopping.features.recommendations.ui.components.StoreRecommendationSummaryItem
-import ke.co.xently.shopping.ui.theme.XentlyTheme
 import kotlin.random.Random
 
 @Composable
