+++ app/src/main/java/ke/co/xently/shopping/features/recommendations/ui/components/HitItem.kt
@@ -20,8 +20,8 @@ import androidx.compose.ui.unit.dp
 import ke.co.xently.shopping.R
 import ke.co.xently.shopping.features.core.currencyNumberFormat
 import ke.co.xently.shopping.features.core.toSystemDefaultZonedDateTime
+import ke.co.xently.shopping.features.core.ui.theme.XentlyTheme
 import ke.co.xently.shopping.features.recommendations.models.Recommendation
-import ke.co.xently.shopping.ui.theme.XentlyTheme
 import java.time.LocalDateTime
 import java.time.ZoneOffset
 import java.time.format.DateTimeFormatter
