+++ app/src/main/java/ke/co/xently/shopping/features/recommendations/ui/components/StoreRecommendationSummaryItem.kt
@@ -24,8 +24,8 @@ import androidx.compose.ui.tooling.preview.Preview
 import androidx.compose.ui.unit.dp
 import ke.co.xently.shopping.R
 import ke.co.xently.shopping.features.core.currencyNumberFormat
+import ke.co.xently.shopping.features.core.ui.theme.XentlyTheme
 import ke.co.xently.shopping.features.recommendations.models.Recommendation
-import ke.co.xently.shopping.ui.theme.XentlyTheme
 import kotlin.random.Random
 
 @Composable
